run data tests after the transformations
