+++ migration_steps/shared/config.py
@@ -11,6 +11,7 @@ class BaseConfig:
     VERBOSE = 5
     DATA = 2
     verbosity_levels = {0: "INFO", 1: "DEBUG", 2: "VERBOSE"}
+    SAMPLE_PERCENTAGE = 1
 
     def verbose(self, msg, *args, **kwargs):
         if logging.getLogger().isEnabledFor(self.VERBOSE):
