+++ migration_steps/transform_casrec/app/data_tests/test_calculated_values.py
@@ -9,31 +9,30 @@ from data_tests.helpers import get_data_from_query
 
 
 @parametrize_with_cases(
-    ("defaults", "source_query", "module_name"),
+    ("calculated_fields", "source_query", "module_name"),
     cases=list_of_test_cases,
-    has_tag="default",
+    has_tag="calculated",
 )
-def test_default_values(get_config, defaults, source_query, module_name):
+def test_default_values(test_config, calculated_fields, source_query, module_name):
     print(f"module_name: {module_name}")
+
     # print(f"source_query: {source_query}")
 
-    config = get_config
+    config = test_config
 
     add_to_tested_list(
-        module_name=module_name, tested_fields=[x for x in defaults.keys()]
+        module_name=module_name, tested_fields=[x for x in calculated_fields.keys()]
     )
 
     source_sample_df = get_data_from_query(
         query=source_query, config=config, sample=True
     )
 
-    # print(source_sample_df.to_markdown())
-
     assert source_sample_df.shape[0] > 0
 
     print(f"Checking {source_sample_df.shape[0]} rows of data ({SAMPLE_PERCENTAGE}%) ")
     assert source_sample_df.shape[0] > 0
-    for k, v in defaults.items():
+    for k, v in calculated_fields.items():
         matches = source_sample_df[k].str.contains(str(v))
         total_matches = matches.sum()
         success = total_matches == source_sample_df.shape[0]
