--- migration_steps/transform_casrec/app/shared_utilities/__init__.py
@@ -1,7 +1,0 @@
-from .helpful_things import helpful2
-from .config import get_config
-from . import custom_logger
-
-
-def helpful():
-    print("I AM VERY HELPFUL local")
