--- migration_steps/transform_casrec/app/shared_utilities/helpful_things.py
@@ -1,2 +1,0 @@
-def helpful2():
-    print("I AM VERY HELPFUL TOO local")
